% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blz2b22.R
\name{blz2b22}
\alias{blz2b22}
\title{B22 in Experimental Unit, 10e-4 mol.ml/g/g}
\usage{
blz2b22(blz, vol, dx, mass)
}
\arguments{
\item{blz}{sum from Boltzmann integartion}

\item{vol}{volume of whole integration space}

\item{dx}{gridspacing}

\item{mass}{mol mass}
}
\description{
B22 in Experimental Unit, 10e-4 mol.ml/g/g
}
\examples{
blz2b22(0.0,32/3*pi*27.241659^3,1.0,66391.13)  #0.4627858 BSA, rg=27.241659^3, mass=66391.13
}

