% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blz2b22.R
\name{blz2b22M}
\alias{blz2b22M}
\title{B22 in Standard Unit M}
\usage{
blz2b22M(blz, vol, dx)
}
\arguments{
\item{blz}{sum from Boltzmann integartion}

\item{vol}{volume of whole integration space}

\item{dx}{gridspacing}
}
\description{
B22 in Standard Unit M
}
\examples{
blz2b22M(0.0,32/3*pi*27.241659^3,1.0) #203.9859, BSA, rg=27.241659^3
}

