% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blz2b22.R
\name{blz2KAMayerdx}
\alias{blz2KAMayerdx}
\title{Convert Boltzmann Integartion into KA}
\usage{
blz2KAMayerdx(blz, vol, dx)
}
\arguments{
\item{blz}{sum from Boltzmann integartion}

\item{vol}{volume of whole integration space}

\item{dx}{gridspacing}
}
\description{
Convert Boltzmann Integartion into KA
}

