% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elec.R
\name{DebyeHuckel}
\alias{DebyeHuckel}
\title{Potential Function for electrostatics in form of Debye–Huckel}
\usage{
DebyeHuckel(r, q1, q2, kap)
}
\arguments{
\item{r}{distance from center}

\item{q1}{charges of A}

\item{q2}{charges of B}

\item{kap}{kappa, 1/lamda, the Debye length}
}
\description{
Potential Function for electrostatics in form of Debye–Huckel
}
\examples{
hardsphere(0.5, 1.0)
hardsphere(1.5, 1.0)
}

