% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glj.R
\name{glj}
\alias{glj}
\title{Potential Function for Generized Lennard–Jones}
\usage{
glj(r, eps, sig, offset)
}
\arguments{
\item{r}{distance from center}

\item{eps}{epsilon, strength}

\item{sig}{sigma, size of sphere}

\item{offset, }{offset of sphere center}
}
\description{
Potential Function for Generized Lennard–Jones
}
\examples{
glj(0.5, 1.0, 1.0, 0.1)
glj(1.5, 1.0, 1.0, 0.1)
}
\keyword{Generized}
\keyword{Lennard-Jones}
\keyword{glj}

