% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glj.R
\name{gLJCut}
\alias{gLJCut}
\title{Potential Function for Generized Lennard–Jones with cutoff}
\usage{
gLJCut(r, eps, sig, offset, rcut)
}
\arguments{
\item{r}{distance from center}

\item{eps}{epsilon, strength}

\item{sig}{sigma, size of sphere}

\item{offset, }{offset of sphere center}

\item{rcut}{cutoff}
}
\description{
Potential Function for Generized Lennard–Jones with cutoff
}
\examples{
gLJCut(0.5, 1.0, 1.0, 0.1, 2.5)
gLJCut(1.5, 1.0, 1.0, 0.1, 2.5)
}
\keyword{Generized}
\keyword{Lennard-Jones}
\keyword{glj}

