% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glj.R
\name{gljforce}
\alias{gljforce}
\title{Force Function for Generized Lennard–Jones}
\usage{
gljforce(r, eps, sig, offset)
}
\arguments{
\item{r}{distance from center}

\item{eps}{epsilon, strength}

\item{sig}{sigma, size of sphere}

\item{offset, }{offset of sphere center}
}
\description{
Force Function for Generized Lennard–Jones
}

