% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glj.R
\name{gLJShf}
\alias{gLJShf}
\title{Potential Function for Generized Lennard–Jones with Linear-Force shifted}
\usage{
gLJShf(r, eps, sig, offset, rcut)
}
\arguments{
\item{r}{distance from center}

\item{eps}{epsilon, strength}

\item{sig}{sigma, size of sphere}

\item{rcut}{cutoff}
}
\description{
Potential Function for Generized Lennard–Jones with Linear-Force shifted
}
\examples{
gLJShf(0.5, 1.0, 1.0, 0.1, 2.5)
gLJShf(1.5, 1.0, 1.0, 0.1, 2.5)
}
\keyword{Generized}
\keyword{Lennard-Jones}
\keyword{glj}

