% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardsphere.R
\name{HardSph}
\alias{HardSph}
\title{Wrapper of Potential Function for Hard Sphere}
\usage{
HardSph(r, eps, sig, rcut)
}
\arguments{
\item{r}{distance from center}

\item{eps}{epsilon, ignored}

\item{sig}{sigma, size of sphere}

\item{rcut}{cutoff, ignored}

\item{pot}{potential}
}
\description{
This is a wrapper over hardsphere to called by SphB22
}
\examples{
HardSph(0.5, 1.0, 1.0, 100)
HardSph(1.5, 1.0, 1.0, 100)
}
\keyword{hardsphere}

