% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardsphere.R
\name{hardsphere}
\alias{hardsphere}
\title{Potential Function for Hard Sphere}
\usage{
hardsphere(r, sig)
}
\arguments{
\item{r}{distance from center}

\item{sig}{sigma, size of sphere}
}
\description{
Potential Function for Hard Sphere
}
\examples{
hardsphere(0.5, 1.0)
hardsphere(1.5, 1.0)
}
\keyword{hardsphere}

