% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lennardjones.R
\name{lj}
\alias{lj}
\title{Potential Function for Lennard–Jones}
\usage{
lj(r, eps, sig)
}
\arguments{
\item{r}{distance from center}

\item{eps}{epsilon, strength}

\item{sig}{sigma, size of sphere}
}
\description{
Potential Function for Lennard–Jones
}
\examples{
lj(0.5, 1.0, 1.0)
lj(1.5, 1.0, 1.0)
}
\keyword{Lennard-Jones}
\keyword{lj}

