% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lennardjones.R
\name{LJCut}
\alias{LJCut}
\title{Potential Function for Lennard–Jones with cutoff}
\usage{
LJCut(r, eps, sig, rcut)
}
\arguments{
\item{r}{distance from center}

\item{eps}{epsilon, strength}

\item{sig}{sigma, size of sphere}

\item{rcut}{cutoff}
}
\description{
Potential Function for Lennard–Jones with cutoff
}
\examples{
LJCut(0.5, 1.0, 1.0, 2.5)
LJCut(1.5, 1.0, 1.0, 2.5)
}
\keyword{Lennard-Jones}
\keyword{lj}

