% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lennardjones.R
\name{ljforce}
\alias{ljforce}
\title{Force Function for Lennard–Jones}
\usage{
ljforce(r, eps, sig)
}
\arguments{
\item{r}{distance from center}

\item{eps}{epsilon, strength}

\item{sig}{sigma, size of sphere}
}
\description{
Force Function for Lennard–Jones
}

