% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphb22.R
\name{SphB22}
\alias{SphB22}
\title{B22 for sphere with r depended potential, J Biol Phys 2015 41:85-97 eq.3.}
\usage{
SphB22(pot, beta, ..., lower = 0, upper = Inf)
}
\arguments{
\item{pot}{potential function in form of pot(r, ...)}

\item{beta}{kB/T}

\item{lower}{lower limit for integration}

\item{upper}{upper limit for integration}
}
\description{
B22 for sphere with r depended potential, J Biol Phys 2015 41:85-97 eq.3.
}
\examples{
HSB2 <- SphB22(hardsphere, 1, 1, lower = 0, upper = Inf)
cat('HS B2:', HSB2, '\\n')
}

