% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lennardjones.R
\name{LJ}
\alias{LJ}
\title{Wrapper of Potential Function for Lennard–Jones}
\usage{
LJ(r, eps, sig, rcut)
}
\arguments{
\item{r}{distance from center}

\item{eps}{epsilon, strength}

\item{sig}{sigma, size of sphere}

\item{rcut}{cutoff, ignored}
}
\description{
This is a wrapper over lj to called by SphB22
}
\examples{
LJ(0.5, 1.0, 1.0, 2.5)
LJ(1.5, 1.0, 1.0, 2.5)
}
\keyword{Lennard-Jones}
\keyword{lj}

