% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphb22.R
\name{SphB22}
\alias{SphB22}
\title{B22 for sphere with r depended potential, J Biol Phys 2015 41:85-97 eq.3.}
\usage{
SphB22(pot, eps, sig, rcut, beta, lower = 0, upper = Inf)
}
\arguments{
\item{pot}{potential}

\item{eps}{epsilon}

\item{sig}{sigma}

\item{rcut}{cutoff}

\item{beta}{kB/T}

\item{lower}{lower limit for integration}

\item{upper}{upper limit for integration}
}
\description{
B22 for sphere with r depended potential, J Biol Phys 2015 41:85-97 eq.3.
}
\examples{
HSB2 <- SphB22(HardSph, 1, 1, Inf, 1, 0,  Inf)
cat('HS B2:', HSB2, '\\n')
}

