% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphb22.R
\name{u2pc}
\alias{u2pc}
\title{Convert u into Partition for Integration}
\usage{
u2pc(pot)
}
\arguments{
\item{pot}{potential function in form of pot(r, ...)}

\item{beta}{kB/T}
}
\description{
Convert u into Partition for Integration, (exp(-u)*b-1)*x^2, use closures
following http://adv-r.had.co.nz/Functional-programming.html
}
\examples{
partHS <- u2pc(hardsphere)
partHS(0.5, 1.0, 1)
partHS(1.5, 1.0, 1)
}

